+++ app/src/main/java/com/example/compose_magazin/data/api/RetrofitClient.kt
@@ -1,4 +1,4 @@
-package com.example.compose_magazin.dataAccess.api
+package com.example.compose_magazin.data.api
 
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import kotlinx.serialization.ExperimentalSerializationApi
