+++ app/src/main/java/com/example/compose_magazin/data/mappers/Mappers.kt
@@ -1,8 +1,8 @@
-package com.example.compose_magazin.dataAccess.mappers
+package com.example.compose_magazin.data.mappers
 
-import com.example.compose_magazin.dataAccess.models.PetCategoryData
-import com.example.compose_magazin.dataAccess.models.PetProductData
-import com.example.compose_magazin.dataAccess.models.PetTagData
+import com.example.compose_magazin.data.models.PetCategoryData
+import com.example.compose_magazin.data.models.PetProductData
+import com.example.compose_magazin.data.models.PetTagData
 import com.example.compose_magazin.domain.entity.PetCategory
 import com.example.compose_magazin.domain.entity.PetProduct
 import com.example.compose_magazin.domain.entity.PetTag
