+++ app/src/main/java/com/example/compose_magazin/data/mock/MockData.kt
@@ -1,4 +1,4 @@
-package com.example.compose_magazin.dataAccess.mock
+package com.example.compose_magazin.data.mock
 
 import com.example.compose_magazin.domain.entity.PetCategory
 import com.example.compose_magazin.domain.entity.PetProduct
