+++ app/src/main/java/com/example/compose_magazin/di/AppModule.kt
@@ -1,6 +1,6 @@
 package com.example.compose_magazin.di
 
-import com.example.compose_magazin.dataAccess.PetRepositoryImpl
+import com.example.compose_magazin.data.PetRepositoryImpl
 import com.example.compose_magazin.domain.PetRepository
 import dagger.Binds
 import dagger.Module
