+++ app/src/main/java/com/example/compose_magazin/presentation/cart/CartItemCard.kt
@@ -1,4 +1,4 @@
-package com.example.compose_magazin.presentation.cartScreen
+package com.example.compose_magazin.presentation.cart
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -31,7 +31,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.compose_magazin.domain.entity.PetProduct
-import com.example.compose_magazin.presentation.uiComponents.ProductImage
+import com.example.compose_magazin.presentation.components.ProductImage
 
 @Composable
 fun CartProductItem(
