+++ app/src/main/java/com/example/compose_magazin/presentation/scaffold/ScaffoldViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.compose_magazin.presentation.scaffoldComponents
+package com.example.compose_magazin.presentation.scaffold
 
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
@@ -11,10 +11,10 @@ class ScaffoldViewModel @Inject constructor() : ViewModel() {
     private var _cartItemCount = mutableStateOf(0)
     val cartItemCount: MutableState<Int> get() = _cartItemCount
 
-    fun addItem(int: Int) {
-        _cartItemCount.value += int
+    fun addItem(changeValue: Int) {
+        _cartItemCount.value += changeValue
     }
-    fun removeItem(int: Int) {
-        _cartItemCount.value -= int
+    fun removeItem(changeValue: Int) {
+        _cartItemCount.value -= changeValue
     }
 }
